Refactor feature gates into its own package
With some of the changes regarding conformance image registry, sonobuoy
default registry, etc we should extract this into a separate package to
avoid import cycles.
